# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the modular ServeoAPI management server.",
        "title": "ServeoAPI",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v2/docker/containers/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all docker containers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker-containers"
                ],
                "summary": "List Docker Containers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docker.ContainerInfo"
                            }
                        }
                    }
                }
            }
        },
        "/v2/docker/containers/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deploys a new container. Bind mounts are strictly restricted for security.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker-containers"
                ],
                "summary": "Create and Start a Docker Container",
                "parameters": [
                    {
                        "description": "Container configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docker.CreateContainerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docker.ContainerInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/docker/containers/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns detailed information about a container",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker-containers"
                ],
                "summary": "Inspect a Docker Container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a container. Pass ?force=true to force removal of a running container.",
                "tags": [
                    "docker-containers"
                ],
                "summary": "Delete a Docker Container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Force remove",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/v2/docker/containers/{id}/logs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Streams logs of a container using Server-Sent Events",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "docker-containers"
                ],
                "summary": "Stream Container Logs (SSE)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event Stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/docker/containers/{id}/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Streams CPU/RAM stats of a container using Server-Sent Events",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "docker-containers"
                ],
                "summary": "Stream Container Stats (SSE)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event Stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/docker/containers/{id}/{action}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Action must be one of: start, stop, restart",
                "tags": [
                    "docker-containers"
                ],
                "summary": "Perform action on a container (start, stop, restart)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Container ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Action to perform (start, stop, restart)",
                        "name": "action",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docker.ContainerInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "docker.CreateContainerRequest": {
            "type": "object",
            "properties": {
                "env": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ports": {
                    "description": "e.g. {\"8080\": \"80\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "restart_policy": {
                    "type": "string"
                },
                "volumes": {
                    "description": "e.g. [\"myvol:/data\", \"/var/serveoapi/data/app:/app\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}